requirements.txt
numpy==1.22.1
pandas==1.3.5
Pillow==9.0.1
pytz==2019.3
PyYAML==6.0
requests==2.27.1
rpy2==3.4.5
